[AS7-4286] Fix JavaCC grammars for version 5.
The JBossQL en EJBQL grammars used for CMP use some tokens with
empty regular expressions. This worked with JavaCC 4 but doesn't
with version 5.

This patch creates a DUMMY lexical state that will never be
reached and puts those tokens inside with a non empty regular
expression (that will never be matched).